Move pyenvLocator.ts
