+++ src/client/pythonEnvironments/index.ts
@@ -22,7 +22,7 @@ import { CustomVirtualEnvironmentLocator } from './base/locators/lowLevel/custom
 import { CondaEnvironmentLocator } from './base/locators/lowLevel/condaLocator';
 import { GlobalVirtualEnvironmentLocator } from './base/locators/lowLevel/globalVirtualEnvronmentLocator';
 import { PosixKnownPathsLocator } from './base/locators/lowLevel/posixKnownPathsLocator';
-import { PyenvLocator } from './discovery/locators/services/pyenvLocator';
+import { PyenvLocator } from './base/locators/lowLevel/pyenvLocator';
 import { WindowsRegistryLocator } from './discovery/locators/services/windowsRegistryLocator';
 import { WindowsStoreLocator } from './discovery/locators/services/windowsStoreLocator';
 import { getEnvironmentInfoService } from './info/environmentInfoService';
